GUI frame created

+++ ReflectionAPI/src/ie/gmit/sw/Metric.java
@@ -46,7 +46,7 @@ public class Metric {
 		//Make sure outCounter is not zero to avoid division by 0,
 		//inCounter does not need to be checked at this point
 		if (outCounter > 0){
-			stability = (float) outCounter/(inCounter + outCounter);
+			stability = (float) ((float) outCounter/((float)inCounter + (float)outCounter));
 		}
 		else{
 			stability = 0f;
